requirements.txt
fastapi
uvicorn
selenium == 4.0.0.b4
aiohttp
bs4
lxml
requests
deta
line-bot-sdk